requirements.txt
kafka-python
pandas
kagglehub